feat(03-02): create consumer lifecycle tests and decompose test_messenger_reliability
- Move 4 tests from test_messenger_reliability.py to test_corr_consumer.py
- Add new test_consumer_lifecycle_full for complete state machine verification
- Delete fully decomposed test_messenger_reliability.py
